✅ ✨ 🏗️ Architectural changes for game components. Add of opponentPad logic.
